refactor: split media providers
