+++ auth-portal/go.mod
@@ -3,7 +3,7 @@ module auth-portal
 go 1.23.8
 
 require (
-    github.com/golang-jwt/jwt/v5 v5.2.2
-    github.com/gorilla/mux v1.8.0
-    github.com/lib/pq v1.10.9
+	github.com/golang-jwt/jwt/v5 v5.2.2
+	github.com/gorilla/mux v1.8.0
+	github.com/lib/pq v1.10.9
 )
